optimization
optmize code (remove redunduncy) + optimization for round display and black and white display => remove aplite support

+++ src/c/data.h
@@ -14,7 +14,7 @@ char* serveNames[NUM_COMPLEX];
 
 int currentSetIndex;
 
-int8_t dataVariable[NUM_MAX_SET+1][NUM_SETTER_POSITION][NUM_COMPLEX][2];
+int dataVariable[NUM_MAX_SET+1][NUM_SETTER_POSITION][NUM_COMPLEX][2];
 
 int setWon[NUM_MAX_SET];
